[FRF] Add support for a frequency response function
This injects noise from the pseudo-random noise generator into the current
loop. The noise manifests itself on the current control, but in order to
visualize the input we also have to transmit the control input `V_d`. This
commit extends the firmware's telemetry such that it outputs `V_d` at the
same time as it outputs other traced data.